Move OpeningHourNormalizer and Denormalizer to new namespace
