+++ src/Model/Serializer/ValueObject/Calendar/OpeningHours/AdjustedDaysDenormalizer.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpeningHours;
 
 use CultuurNet\UDB3\DateTimeFactory;
-use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpeningHourDenormalizer;
 use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\TranslatedAdjustedDescriptionDenormalizer;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\AdjustedDay;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\AdjustedDays;
