+++ src/Model/Serializer/ValueObject/Calendar/OpeningHours/OpeningHoursNormalizer.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar;
+namespace CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpeningHours;
 
 use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\OpeningHours;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
